python/requirements.txt
spotipy==2.23.0
pyyaml==6.0
tqdm==4.65.0
Flask==2.3.1